fix bug in concatenating vectors
It seems like it used to be possible to combine vectors of varying
length into a 2D numpy array but this now throws an error. We can feed a
list instead of a numpy array into scipy's sparse.diag() which fixes the
bug. The stack overflow article linked in the bug also combines the data
this way.